requirements.txt
absl-py==1.4.0
alabaster==0.7.12
appdirs==1.4.4
arrow==1.2.3
astroid==2.11.7
astunparse==1.6.3
atomicwrites==1.4.0
attrs==22.1.0
audioread==3.0.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
black==22.6.0
bleach==4.1.0
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
clang==5.0
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.5
contourpy==1.0.5
cookiecutter==1.7.3
cryptography==38.0.1
cycler==0.11.0
cytoolz==0.12.0
dask==2022.7.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
docutils==0.18.1
entrypoints==0.4
et-xmlfile==1.1.0
fastjsonschema==2.16.2
flake8==4.0.1
flatbuffers==1.12
flit_core==3.6.0
fonttools==4.25.0
fsspec==2022.11.0
gast==0.4.0
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.1.0
idna==3.4
imagecodecs==2021.8.26
imageio==2.19.3
imagesize==1.4.1
importlib-metadata==4.11.3
importlib-resources==5.2.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.15.2
ipython==7.31.1
ipython-genutils==0.2.0
isort==5.9.3
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.1.1
jsonschema==4.16.0
jupyter_client==7.4.7
jupyter_core==4.11.2
jupyterlab-pygments==0.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
keras-tcn==3.5.0
keyring==23.4.0
kiwisolver==1.4.2
lazy-object-proxy==1.6.0
libclang==15.0.6.1
librosa==0.9.2
llvmlite==0.39.1
locket==1.0.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
mypy-extensions==0.4.3
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.5
networkx==2.8.4
numba==0.56.4
numexpr==2.8.4
numpy==1.23.5
numpydoc==1.5.0
oauthlib==3.2.0
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
paramiko==2.8.1
parso==0.8.3
partd==1.2.0
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==22.3.1
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
plotly==5.9.0
pluggy==1.0.0
ply==3.11
pooch==1.6.0
poyo==0.5.0
prompt-toolkit==3.0.20
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.11.2
pylint==2.14.5
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtWebEngine==5.15.4
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-lsp-black==1.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.5.0
python-slugify==5.0.2
python-version==0.0.2
pytz==2022.1
PyWavelets==1.4.1
pywin32==305.1
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==23.2.0
QDarkStyle==3.0.2
qstylizer==0.1.10
QtAwesome==1.0.3
qtconsole==5.3.2
QtPy==2.2.0
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rope==0.22.0
rsa==4.9
Rtree==0.9.7
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
setuptools==65.5.0
sip==6.6.2
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.11.0
soupsieve==2.3.2.post1
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.3.3
spyder-kernels==2.3.3
tenacity==8.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
text-unidecode==1.3
textdistance==4.2.1
threadpoolctl==2.2.0
three-merge==0.1.1
tifffile==2021.7.2
tinycss==0.4
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
toolz==0.12.0
tornado==6.2
traitlets==5.7.1
typing_extensions==4.4.0
ujson==5.4.0
Unidecode==1.2.0
urllib3==1.26.13
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
whatthepatch==1.0.2
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.14.1
xlrd==2.0.1
yapf==0.31.0
zipp==3.8.0
